Fixes and renamed deprecation to warnings
